4/requirements.txt
geopandas==0.11.1
matplotlib==3.5.0
numpy==1.22.4
opencv_python==4.6.0.66
pandas==1.3.5
Pillow==9.2.0
rasterio==1.2.10
Shapely==1.8.4
torch==1.12.1
torchvision==0.13.1
git+https://github.com/mkisantal/backboned-unet